Add Junit tests
