--- src/main/duke/instruction/Quit.java
@@ -1,31 +1,0 @@
-package main.duke.instruction;
-
-import main.duke.task.TaskManager;
-import main.duke.ui.Ui;
-
-/**
- * Represents the instruction 'quit'.
- */
-final class Quit extends Instruction {
-
-    private static final String SEE_YOU_MESSAGE = "Bye!";
-
-    /**
-     * Constructor, sets the description of the instruction.
-     *
-     * @param message The terminating command that signals the end of program.
-     */
-    Quit(String message, TaskManager tasks) {
-        super(message, tasks);
-    }
-
-    /**
-     * Returns a see-you message.
-     *
-     * @return The see-you message.
-     */
-    @Override
-    public void act(Ui ui) {
-        ui.printMessage(Quit.SEE_YOU_MESSAGE);
-    }
-}
