--- src/main/duke/task/FailedToInterpretTaskException.java
@@ -1,11 +1,0 @@
-package main.duke.task;
-
-import main.duke.main.DukeException;
-
-class FailedToInterpretTaskException extends DukeException {
-
-
-    protected FailedToInterpretTaskException(String message) {
-        super(message);
-    }
-}
