--- src/main/duke/task/TaskNotFoundException.java
@@ -1,15 +1,0 @@
-package main.duke.task;
-
-import main.duke.main.DukeException;
-
-public final class TaskNotFoundException extends DukeException {
-
-
-    protected TaskNotFoundException(String message) {
-        super(message);
-    }
-
-    protected TaskNotFoundException() {
-        super("Oops, I'm not able to retrieve this task");
-    }
-}
