--- src/main/duke/task/ToDo.java
@@ -1,47 +1,0 @@
-package main.duke.task;
-
-import java.util.HashMap;
-
-/**
- * Represents a to-do task. There is no deadline or timeslot associated.
- */
-final class ToDo extends Task {
-
-    /**
-     * Initializes a to-do item, with a description.
-     *
-     * @param description The description of the task.
-     */
-    protected ToDo(String description) {
-        super(description);
-    }
-
-    protected ToDo(HashMap<String, Object> infoTable) {
-        super(infoTable);
-    }
-
-
-    @Override
-    protected String getTypeIcon() {
-        return "[T]";
-    }
-
-    @Override
-    protected TaskType getType() {
-        return TaskType.TODO;
-    }
-
-
-    /**
-     * Returns a hash map that contains the data to be stored into hard disk.
-     *
-     * @return The hash map containing the data.
-     */
-    @Override
-    public HashMap<String, Object> getInfoTable() {
-
-        var infoTable = this.initializeInfoTable();
-        return infoTable;
-    }
-
-}
