+++ src/main/java/duke/instruction/InvalidInstructionException.java
@@ -1,0 +1,13 @@
+package main.java.duke.instruction;
+
+import main.java.duke.main.DukeException;
+
+/**
+ * Represents the exception that occurs when the user input of command is invalid.
+ */
+public class InvalidInstructionException extends DukeException {
+
+    protected InvalidInstructionException(String message) {
+        super(message);
+    }
+}
