+++ src/main/java/duke/task/FailedToInterpretTaskException.java
@@ -1,0 +1,11 @@
+package main.java.duke.task;
+
+import main.java.duke.main.DukeException;
+
+public class FailedToInterpretTaskException extends DukeException {
+
+
+    protected FailedToInterpretTaskException(String message) {
+        super(message);
+    }
+}
