+++ src/main/java/duke/task/TaskType.java
@@ -1,0 +1,12 @@
+package main.java.duke.task;
+
+/**
+ * Describes types of tasks.
+ * Currently, there are 3 types of tasks: <code>task.Event</code>, <code>Todo</code>, <code>task.Deadline</code>.
+ */
+public enum TaskType {
+
+    EVENT,
+    TODO,
+    DEADLINE
+}
