+++ src/main/java/duke/ui/UiException.java
@@ -1,0 +1,19 @@
+package main.java.duke.ui;
+
+import java.io.IOException;
+
+/**
+ * Represents exceptions occurring in UI.
+ */
+class UiException extends IOException {
+
+
+    /**
+     * Constructs an <code>ui.UiException</code> with the specified message.
+     *
+     * @param message The message that describes the exception.
+     */
+    UiException(String message) {
+        super(message);
+    }
+}
